Crate anchor_attribute_event

Macros

emit

Logs an event that can be subscribed to by clients. Uses the sol_log_data syscall which results in the following log:

Attribute Macros

event

The event attribute allows a struct to be used with emit! so that programs can log significant events in their programs that clients can subscribe to. Currently, this macro is for structs only.